# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SearchPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "title": "Items"
          },
          "next_cursor": {
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "SearchPage",
        "type": "object"
      },
      "SkillCreate": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "namespace": {
            "title": "Namespace",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "maxItems": 50,
            "title": "Tags",
            "type": "array"
          },
          "visibility": {
            "default": "private",
            "enum": [
              "private",
              "public"
            ],
            "title": "Visibility",
            "type": "string"
          }
        },
        "required": [
          "namespace",
          "name",
          "description"
        ],
        "title": "SkillCreate",
        "type": "object"
      },
      "SkillDetail": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "title": "Description"
          },
          "id": {
            "title": "Id"
          },
          "label_details": {
            "title": "Label Details"
          },
          "labels": {
            "title": "Labels"
          },
          "name": {
            "title": "Name"
          },
          "namespace": {
            "title": "Namespace"
          },
          "owner_id": {
            "title": "Owner Id"
          },
          "tags": {
            "title": "Tags"
          },
          "updated_at": {
            "title": "Updated At"
          },
          "versions": {
            "title": "Versions"
          },
          "visibility": {
            "title": "Visibility"
          }
        },
        "required": [
          "id",
          "namespace",
          "name",
          "description",
          "visibility",
          "owner_id",
          "tags",
          "updated_at",
          "versions",
          "label_details"
        ],
        "title": "SkillDetail",
        "type": "object"
      },
      "SkillSummary": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "title": "Description"
          },
          "id": {
            "title": "Id"
          },
          "labels": {
            "title": "Labels"
          },
          "name": {
            "title": "Name"
          },
          "namespace": {
            "title": "Namespace"
          },
          "owner_id": {
            "title": "Owner Id"
          },
          "tags": {
            "title": "Tags"
          },
          "updated_at": {
            "title": "Updated At"
          },
          "visibility": {
            "title": "Visibility"
          }
        },
        "required": [
          "id",
          "namespace",
          "name",
          "description",
          "visibility",
          "owner_id",
          "tags",
          "updated_at"
        ],
        "title": "SkillSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Private-first immutable Agent Skill registry. Errors use RFC 9457 Problem Details.",
    "title": "Personal Skill Hub Registry API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/skills": {
      "get": {
        "operationId": "list_skills_api_v1_skills_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 256,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "type": "string"
              },
              "title": "Tag",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "compatibility",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Compatibility"
            }
          },
          {
            "in": "query",
            "name": "owner",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 255,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2048,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Skills",
        "tags": [
          "registry"
        ]
      },
      "post": {
        "operationId": "create_skill_api_v1_skills_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 255,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Skill",
        "tags": [
          "registry"
        ]
      }
    },
    "/api/v1/skills/{namespace}/{name}": {
      "get": {
        "operationId": "get_skill_api_v1_skills__namespace___name__get",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "required": true,
            "schema": {
              "title": "Namespace",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Skill",
        "tags": [
          "registry"
        ]
      }
    },
...
  }
}